refactor: relocate src/gateway → src/legacy/gateway pending refactor
