+++ src/legacy/plugins/discord/index.ts
@@ -1,6 +1,6 @@
-import type { Transport } from "../../../gateway/types.js";
+import type { Transport } from "../../../legacy/gateway/types.js";
 import type { SessionStore } from "../../../sessions/types.js";
-import type { LazyTransportContext } from "../../../gateway/transport-context.js";
+import type { LazyTransportContext } from "../../../legacy/gateway/transport-context.js";
 import type { SessionStoreManager } from "../../../agent/runners/pi/session-store.js";
 import type { AgentRuntime } from "../../../agent/runtime.js";
 import type { IsotopesConfigFile } from "../../../config.js";
